Refactor: Decouple TeamMembers page logic and extract UI components
